feat: change color if pixivision is viewed feat: need to restart if change use_new_banner fix: add padding & larger minWidth to prevent first item fail to be centered by SnapHelper

+++ app/src/main/java/com/perol/asdpl/pixivez/objects/ScreenUtil.kt
@@ -4,9 +4,8 @@ import android.content.res.Resources
 
 class ScreenUtil {
     companion object {
-
         @JvmStatic
-        fun dip2px(dpValue: Int): Int {
+        fun dip2px(dpValue: Float): Int {
             return (0.5f + dpValue * Resources.getSystem()
                 .displayMetrics.density).toInt()
         }
